refactor: rearrange test related files
